Removing some residual obsolete inputs and overall working on integration
