+++ tests/regression/config.cfg
@@ -10,18 +10,6 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 
-% Web & flange thickness [m]
-W_THICKNESS = 0.02
-
-% web height [m]
-W_HEIGHT = 0.40
-
-% flange width [m]
-F_WIDTH = 0.20
-
-% Number of overall nodes along the wing (no collapsed) (n. FEM elements = N_NODES -1 )  %% TO BE REMOVED
-N_NODES = 21
-
 % Young modulus [GPa]
 Y_MODULUS = 70000000000
 
@@ -30,15 +18,6 @@ POISSON = 0.3
 
 %  Beam Density [kg/m^3]
 RHO = 2700
-
-% BEAM Length [m]
-B_LENGTH = 30
-
-% Number of rigid modes to be calculated
-N_DOF = 6  
-
-% Applied load [N]	
-LOAD = 0
 
 % Follower load flag (0 = NON_FOLLOWER, 1 = FOLLOWER, 2 = APPROX_FOLLOWER) (DEFAULT = 0) (0 implemented)
 FOLLOWER_FLAG = 0
